Split bn_quokka export module into phase packages
export.py had grown to ~1650 lines covering nine concerns. Split it
along the semantic clusters of the schema while keeping
bn_quokka.export as the stable import surface, so the plugin entry
point, the headless CLI, tests, and any external script keep their
imports unchanged (including the historical __all__):

- context.py: ExportContext shared by all phases, ExportCancelled
- exporters/binary.py: program image (Meta, Segment, Layout, Data)
- exporters/types.py: type table, type xrefs, header collection
- exporters/cfg.py: functions, basic-block splitting, edges
- exporters/instructions.py: instruction/operand token machinery
- exporters/references.py: cross-reference export and classification
- export.py: pipeline orchestration and entry points only

The import graph is a DAG (util <- context <- exporters <- export);
helpers shared across phases became public in their owning module
(extract_mnemonic, token_value, ... in instructions;
map_calling_convention moved to util since both meta and cfg need it).
All code moved verbatim apart from those visibility changes.

README documents the new layout.
