+++ binaryninja_extension/bn_quokka/__init__.py
@@ -1,7 +1,8 @@
-from .export import export_binary_view, export_file, run_export_pipeline
+from .export import ExportCancelled, export_binary_view, export_file, run_export_pipeline
 from .quokka_pb2 import Quokka
 
 __all__ = [
+    "ExportCancelled",
     "Quokka",
     "export_binary_view",
     "export_file",
